move function to normalize.py
